+++ src/SnowflakeSync.php
@@ -2,16 +2,19 @@
 
 namespace Bernskiold\LaravelSnowflakeSync;
 
+use Bernskiold\LaravelSnowflakeSync\Jobs\FlushSnowflakeSync;
 use Bernskiold\LaravelSnowflakeSync\Jobs\RemoveFromSnowflake;
 use Bernskiold\LaravelSnowflakeSync\Jobs\SnowflakeImport;
 
 class SnowflakeSync
 {
-    public const VERSION = '1.0.0';
+    public const VERSION = '0.3.0';
 
     public static $importJob = SnowflakeImport::class;
 
     public static $removeJob = RemoveFromSnowflake::class;
+
+    public static $flushJob = FlushSnowflakeSync::class;
 
     public static $engine = SnowflakeEngine::class;
 
@@ -23,6 +26,11 @@ class SnowflakeSync
     public static function removeUsing(string $class): void
     {
         static::$removeJob = $class;
+    }
+
+    public static function flushUsing(string $class): void
+    {
+        static::$flushJob = $class;
     }
 
     public static function engine(string $class): void